removed 'app' module and moved business logic to 'domain'
