+++ schedule/domain/src/main/java/com/github/schedule/workmonth/WorkMonth.java
