+++ schedule/domain/src/main/java/com/github/schedule/workmonth/event/WorkMonthCreatedEvent.java
