+++ schedule/domain/src/main/java/com/github/schedule/workmonth/exception/WorkMonthExistsException.java
