+++ schedule/domain/src/test/java/com/github/schedule/workmonth/vo/WorkDayTest.java
